src/backend/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
pydantic==2.12.5

# Data
h5py
numpy

# Embeddings
google-generativeai

# Vector DB
qdrant-client

# PDF extraction
pypdf

# Utilities
python-dotenv
python-multipart
